flask/requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
filelock==3.16.1
fsspec==2024.9.0
flask==2.3.3
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai-whisper @ git+https://github.com/openai/whisper.git@279133e3107392276dc509148da1f41bfb532c7e
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.1.0
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
# triton==3.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
gunicorn==23.0.0
yt-dlp==2025.2.19
groq
flask-cors
python-dotenv